Move Group model into h.models package
As part of an effort to move all models into the h.models package, this
moves the Group model out of h.groups and into h.models.
